Обновлены модели. Добавлены скидки
