+++ DiscountsAndPromotionsApp/Network/ProductNetworkServices/Models/ProductResponseModel.swift
@@ -6,14 +6,15 @@ struct ProductResponseModel: Codable {
     let rating: Int?
     let category: CategoryResponseModel?
     let description: String?
-    let image: Int?
-//    нужна правка модели
-//    let stores: [StoreElementResponseModel]?
+    let mainImage: String?
+    let barcode: String
+    let stores: ProductsInStore?
     let isFavorited: Bool
 
     enum CodingKeys: String, CodingKey {
-        case id, name, rating, category, description, image
+        case id, name, rating, category, description, barcode, stores
         case isFavorited = "is_favorited"
+        case mainImage = "main_image"
     }
 }
