+++ DiscountsAndPromotionsApp/Network/ProductNetworkServices/Models/StoreResponseModel.swift
@@ -1,5 +1,14 @@
 import Foundation
 
+struct StoreElementResponseModel: Codable {
+   let id: Int
+   let discount: DiscountResponseModel?
+ //  let store: StoreResponseModel?
+   let price: String?
+}
+
+typealias ProductsInStore = [StoreElementResponseModel]
+
 struct StoreResponseModel: Codable {
    let id: Int?
    let location: LocationResponseModel?
@@ -13,30 +22,9 @@ struct StoreResponseModel: Codable {
    }
 }
 
-struct StoreElementResponseModel: Codable {
-   let id: Int?
-   let discount: Discount?
-   let store: StoreResponseModel?
-   let price: String?
-}
-
 struct LocationResponseModel: Codable {
    let id: Int
    let region, city, street, building: String?
-}
-
-struct DiscountResponseModel: Codable {
-   let discountRate: Int
-   let discountUnit, discountStart, discountEnd: String?
-   let discountCard: Bool
-
-   enum CodingKeys: String, CodingKey {
-       case discountRate = "discount_rate"
-       case discountUnit = "discount_unit"
-       case discountStart = "discount_start"
-       case discountEnd = "discount_end"
-       case discountCard = "discount_card"
-   }
 }
 
 struct StoreChainResponseModel: Codable {
